+++ frontend/app/components/public-schools/existing-conditions/table-existing-conditions.js
@@ -22,9 +22,9 @@ export default Component.extend({
 
   table: computed('activeSdId', 'activeSchoolsLevel', function() {
     if (this.activeSchoolsLevel === 'hs') {
-      return this.analysis.schoolTotals.findBy('level', 'hs');
+      return this.analysis.aggregateTotals.findBy('level', 'hs');
     } else {
-      return this.analysis.schoolTotals.find(
+      return this.analysis.aggregateTotals.find(
         (total) => (total.id === parseInt(this.activeSdId) && total.level === this.activeSchoolsLevel)
       );
     }
