+++ frontend/app/transforms/public-schools/residential-developments.js
@@ -2,7 +2,7 @@ import DS from 'ember-data';
 const { Transform } = DS;
 import FutureResidentialDevelopment from '../../fragments/public-schools/FutureResidentialDevelopment';
 
-export default class PublicSchoolsResidentialDevelopmentTransform extends Transform {
+export default class PublicSchoolsResidentialDevelopmentsTransform extends Transform {
   deserialize(serialized) {        
     return serialized.map((b) => {
       // Defense from saved totals
