+++ frontend/app/transforms/public-schools/schools.js
@@ -1,9 +1,9 @@
 import DS from 'ember-data';
-import Building from '../fragments/public-schools/Building';
+import School from '../../fragments/public-schools/School';
 
 export default DS.Transform.extend({
   deserialize(serialized) {
-    return serialized.map(b => Building.create(b));
+    return serialized.map(b => School.create(b));
   },
 
   serialize(deserialized) {
